LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__host.header.https-expected.txt


PASS __Host: Secure origin: Does not set 'Path=/;'
FAIL __HoSt: Secure origin: Does not set 'Path=/;' assert_equals: expected (undefined) undefined but got (string) "bar2"
PASS __Host: Secure origin: Does set 'Secure; Path=/;'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; '
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; ' assert_equals: expected (undefined) undefined but got (string) "bar6"
PASS __Host: Secure origin: Does not set 'Path=/;MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "bar8"
PASS __Host: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "bar12"
PASS __Host: Secure origin: Does not set 'Path=/;HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Path=/;HttpOnly' assert_equals: expected (undefined) undefined but got (string) "bar14"
PASS __Host: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; HttpOnly' assert_equals: expected (undefined) undefined but got (string) "bar18"
PASS __Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py' assert_equals: expected (undefined) undefined but got (string) "bar20"
